chore: move sb-locale-switcher to helpers
