+++ packages/helpers/index.js
@@ -3,3 +3,4 @@ export { renderLitAsNode } from './renderLitAsNode/src/renderLitAsNode.js';
 
 // Components
 export { SbActionLogger } from './sb-action-logger/src/SbActionLogger.js';
+export { SbLocaleSwitcher } from './sb-locale-switcher/src/SbLocaleSwitcher.js';
