+++ packages/helpers/sb-locale-switcher.js
@@ -1,0 +1,3 @@
+import { SbLocaleSwitcher } from './sb-locale-switcher/src/SbLocaleSwitcher.js';
+
+window.customElements.define('sb-locale-switcher', SbLocaleSwitcher);
