+++ packages/localize/package.json
@@ -38,6 +38,7 @@
   },
   "devDependencies": {
     "@bundled-es-modules/fetch-mock": "^6.5.2",
+    "@lion/helpers": "0.1.0",
     "@open-wc/demoing-storybook": "^1.3.0",
     "@open-wc/testing": "^2.3.4",
     "sinon": "^7.2.2"
